move bowserlib to bokeh.util.browser
