+++ examples/embed/embed_responsive_width_height.py
@@ -9,7 +9,7 @@ body element in this case. A more realistic example might be embedding
 a plot in a Phosphor widget.
 """
 
-from bokeh.browserlib import view
+from bokeh.util.browser import view
 from bokeh.plotting import figure
 from bokeh.embed import components
 from bokeh.resources import INLINE
