+++ examples/glyphs/airports_map.py
@@ -2,7 +2,7 @@ from __future__ import print_function
 
 import json
 
-from bokeh.browserlib import view
+from bokeh.util.browser import view
 from bokeh.document import Document
 from bokeh.embed import file_html
 from bokeh.resources import INLINE
