+++ examples/glyphs/colors.py
@@ -8,7 +8,7 @@ from bokeh.models.glyphs import Rect
 from bokeh.document import Document
 from bokeh.embed import file_html
 from bokeh.resources import INLINE
-from bokeh.browserlib import view
+from bokeh.util.browser import view
 
 css3_colors = pd.DataFrame([
     ("Pink",                  "#FFC0CB", "Pink"),
