+++ examples/glyphs/custom.py
@@ -1,6 +1,6 @@
 from __future__ import print_function
 
-from bokeh.browserlib import view
+from bokeh.util.browser import view
 from bokeh.document import Document
 from bokeh.embed import file_html
 from bokeh.properties import String
