+++ examples/glyphs/dateaxis.py
@@ -4,7 +4,7 @@ from numpy import pi, arange, sin
 import numpy as np
 import time
 
-from bokeh.browserlib import view
+from bokeh.util.browser import view
 from bokeh.document import Document
 from bokeh.embed import file_html
 from bokeh.models.glyphs import Circle
