+++ examples/glyphs/daylight.py
@@ -3,7 +3,7 @@ from __future__ import print_function
 import numpy as np
 import datetime as dt
 
-from bokeh.browserlib import view
+from bokeh.util.browser import view
 from bokeh.document import Document
 from bokeh.embed import file_html
 from bokeh.models.glyphs import Patch, Line, Text
