+++ examples/glyphs/gauges.py
@@ -5,7 +5,7 @@ from math import pi, sin, cos
 from bokeh.document import Document
 from bokeh.embed import file_html
 from bokeh.resources import INLINE
-from bokeh.browserlib import view
+from bokeh.util.browser import view
 
 from bokeh.models.glyphs import Circle, Arc, Ray, Text
 from bokeh.models import ColumnDataSource, Range1d, Plot
