+++ examples/glyphs/glyph2_server.py
@@ -2,7 +2,7 @@ from __future__ import print_function
 
 from numpy import pi, arange, sin, cos
 
-from bokeh.browserlib import view
+from bokeh.util.browser import view
 from bokeh.document import Document
 from bokeh.models.glyphs import Circle
 from bokeh.models import (
