+++ examples/glyphs/glyphs.py
@@ -10,7 +10,7 @@ from bokeh.models.glyphs import (
 from bokeh.document import Document
 from bokeh.embed import file_html
 from bokeh.resources import INLINE
-from bokeh.browserlib import view
+from bokeh.util.browser import view
 
 N = 9
