+++ examples/glyphs/image_url.py
@@ -1,7 +1,7 @@
 
 import numpy as np
 
-from bokeh.browserlib import view
+from bokeh.util.browser import view
 from bokeh.document import Document
 from bokeh.embed import file_html
 from bokeh.models.glyphs import ImageURL
