+++ examples/glyphs/iris_splom.py
@@ -2,7 +2,7 @@ from __future__ import print_function
 
 from math import pi
 
-from bokeh.browserlib import view
+from bokeh.util.browser import view
 from bokeh.document import Document
 from bokeh.embed import file_html
 from bokeh.models.glyphs import Circle, Text
