+++ examples/glyphs/maps_cities.py
@@ -1,6 +1,6 @@
 from __future__ import print_function
 
-from bokeh.browserlib import view
+from bokeh.util.browser import view
 from bokeh.embed import file_html
 from bokeh.models.glyphs import Circle
 from bokeh.models import (
