+++ examples/glyphs/population_server.py
@@ -2,7 +2,7 @@ from __future__ import print_function
 
 from math import pi
 
-from bokeh.browserlib import view
+from bokeh.util.browser import view
 from bokeh.document import Document
 from bokeh.models.glyphs import Line, Quad
 from bokeh.models import (
