+++ examples/glyphs/prim_server.py
@@ -2,7 +2,7 @@ from __future__ import print_function
 
 import numpy as np
 
-from bokeh.browserlib import view
+from bokeh.util.browser import view
 from bokeh.document import Document
 from bokeh.models.glyphs import *
 from bokeh.models import (
