+++ examples/glyphs/trail.py
@@ -10,7 +10,7 @@ import scipy.ndimage as im
 from bokeh.document import Document
 from bokeh.embed import file_html
 from bokeh.resources import INLINE
-from bokeh.browserlib import view
+from bokeh.util.browser import view
 
 from bokeh.models.glyphs import Line, Patches
 from bokeh.models.widgets import VBox
